Redistribution of code

+++ testar/src/nl/ou/testar/StateModel/RLModelManager.java
@@ -90,10 +90,10 @@ public class RLModelManager extends ModelManager implements StateModelManager {
             // get reward and Q-value
             float reward = rewardFunction.getReward(state, getCurrentConcreteState(), currentAbstractState, selectedAbstractAction, actions);
             System.out.println("REWARD: " + Float.toString(reward));
-            final double qValue = qFunction.getQValue((Tag<Float>)this.tag, previousAbstractActionToExecute, selectedAbstractAction, reward, currentAbstractState, actions);
+            final float qValue = qFunction.getQValue((Tag<Float>)this.tag, previousAbstractActionToExecute, selectedAbstractAction, reward, currentAbstractState, actions);
 
             // set attribute for saving in the graph database
-            selectedAbstractAction.addAttribute(tag, (float) qValue);
+            selectedAbstractAction.addAttribute(tag, qValue);
 
             // set previousActionUnderExecute to current abstractActionToExecute for the next iteration
             previousAbstractActionToExecute = selectedAbstractAction;
